--- app/src/main/java/edu/com/trailslist/TrailDaoProvider.kt
@@ -1,5 +1,0 @@
-package edu.com.trailslist
-
-interface TrailDaoProvider {
-    fun provideTrailDao(): TrailDao
-}
