+++ app/src/main/java/edu/com/trailslist/database/provider/TrailDaoProvider.kt
@@ -1,0 +1,7 @@
+package edu.com.trailslist.database.provider
+
+import edu.com.trailslist.database.dao.TrailDao
+
+interface TrailDaoProvider {
+    fun provideTrailDao(): TrailDao
+}
